Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/workers/baseurl/alpha/import-in-moduleworker-expected.txt

CONSOLE MESSAGE: line 1: SyntaxError: Unexpected identifier 'r'. import call expects exactly one argument.

FAIL Base URL in module dedicated workers: import assert_unreached: Got error event Reached unreachable code
